refactor: move all js files to js/ directory
