+++ src/static/diary/add/js/run.js
