+++ src/static/diary/add/js/utils.js
