+++ src/templates/diary/add.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Diary</title>
-    <script src="{{ url_for('static', filename='diary/add/run.js') }}" type="module"></script>
+    <script src="{{ url_for('static', filename='diary/add/js/run.js') }}" type="module"></script>
 
     <link href="{{ url_for('static', filename='diary/favicon.ico') }}" rel="shortcut icon" />
